Move sync_hover functions to presentation.rs
